requirements.txt
# Core Dependencies
pandas>=1.3.0
numpy>=1.20.0
matplotlib>=3.5.0
seaborn>=0.11.0
requests>=2.28.2
urllib3<2.0.0  # Compatibility with current libraries

# API Integration & Data Fetching
python-dotenv>=0.19.0
gspread==6.2.0  # Google Sheets integration
oauth2client==4.1.3  # Auth for Google APIs

# Optimization & Analysis
scipy>=1.10.0
statsmodels>=0.14.0
pyDOE>=0.3.8  # Design of Experiments (for Latin hypercube)

# Parallel Processing
tqdm>=4.65.0  # Progress bars for optimization
joblib>=1.2.0  # Parallel computing for backtests
multiprocess>=0.70.12  # Enhanced multiprocessing

# CLI & Formatting
click>=8.1.0  # Command-line interfaces
rich>=13.3.0  # Rich terminal output
tabulate>=0.9.0  # Table formatting

# Scheduling & Automation
schedule>=1.2.0  # For periodic tasks and data updates

# Testing
pytest>=7.3.0

# Data Persistence
h5py>=3.7.0  # Optional: For large dataset storage 